requirements.txt
opencv-python>=4.5.0
cryptography>=3.4.0
Pillow>=8.0.0
numpy>=1.19.0
aiohttp>=3.8.0
solders>=0.9.0
solana>=0.29.0
python-dotenv>=0.19.0
